# Move mesh generation to vispy