--- a/napari/benchmarks/benchmark_qt_viewer_vectors.py
+++ b/napari/benchmarks/benchmark_qt_viewer_vectors.py
@@ -7,6 +7,7 @@ from qtpy.QtWidgets import QApplication
 
 import napari
 
+
 class QtViewerViewVectorSuite:
     """Benchmarks for viewing vectors in the viewer."""