--- a/napari/layers/vectors/_vector_utils.py
+++ b/napari/layers/vectors/_vector_utils.py
@@ -1,10 +1,8 @@
-from copy import copy
 from typing import Optional, Tuple
 
 import numpy as np
 
 from ...utils.translations import trans
-from ..utils.layer_utils import segment_normal
 
 
 def convert_image_to_coordinates(vectors):
@@ -38,88 +36,6 @@ def convert_image_to_coordinates(vectors):
     coords[:, 1, :] = np.reshape(vectors, (-1, vectors.ndim - 1))
 
     return coords
-
-
-def generate_vector_meshes(vectors, width, length):
-    """Generates list of mesh vertices and triangles from a list of vectors
-
-    Parameters
-    ----------
-    vectors : (N, 2, D) array
-        A list of N vectors with start point and projections of the vector
-        in D dimensions, where D is 2 or 3.
-    width : float
-        width of the line to be drawn
-    length : float
-        length multiplier of the line to be drawn
-
-    Returns
-    -------
-    vertices : (4N, D) array
-        Vertices of all triangles for the lines
-    triangles : (2N, 3) array
-        Vertex indices that form the mesh triangles
-    """
-    ndim = vectors.shape[2]
-    if ndim == 2:
-        vertices, triangles = generate_vector_meshes_2D(vectors, width, length)
-    else:
-        v_a, t_a = generate_vector_meshes_2D(
-            vectors, width, length, p=(0, 0, 1)
-        )
-        v_b, t_b = generate_vector_meshes_2D(
-            vectors, width, length, p=(1, 0, 0)
-        )
-        vertices = np.concatenate([v_a, v_b], axis=0)
-        triangles = np.concatenate([t_a, len(v_a) + t_b], axis=0)
-
-    return vertices, triangles
-
-
-def generate_vector_meshes_2D(vectors, width, length, p=(0, 0, 1)):
-    """Generates list of mesh vertices and triangles from a list of vectors
-
-    Parameters
-    ----------
-    vectors : (N, 2, D) array
-        A list of N vectors with start point and projections of the vector
-        in D dimensions, where D is 2 or 3.
-    width : float
-        width of the line to be drawn
-    length : float
-        length multiplier of the line to be drawn
-    p : 3-tuple, optional
-        orthogonal vector for segment calculation in 3D.
-
-    Returns
-    -------
-    vertices : (4N, D) array
-        Vertices of all triangles for the lines
-    triangles : (2N, 3) array
-        Vertex indices that form the mesh triangles
-    """
-    ndim = vectors.shape[2]
-    vectors = np.reshape(copy(vectors), (-1, ndim))
-    vectors[1::2] = vectors[::2] + length * vectors[1::2]
-
-    centers = np.repeat(vectors, 2, axis=0)
-    offsets = segment_normal(vectors[::2, :], vectors[1::2, :], p=p)
-    offsets = np.repeat(offsets, 4, axis=0)
-    signs = np.ones((len(offsets), ndim))
-    signs[::2] = -1
-    offsets = offsets * signs
-
-    vertices = centers + width * offsets / 2
-    triangles = np.array(
-        [
-            [2 * i, 2 * i + 1, 2 * i + 2]
-            if i % 2 == 0
-            else [2 * i - 1, 2 * i, 2 * i + 1]
-            for i in range(len(vectors))
-        ]
-    ).astype(np.uint32)
-
-    return vertices, triangles
 
 
 def fix_data_vectors(