@kcpevey 2022-08-01
@sofroniewn Is this git approach what you're looking for? https://github.com/sofroniewn/napari/pull/17